fix:Unzip file after downloading it
