+++ sdstore/sdstore_test.go
@@ -1,7 +1,6 @@
 package sdstore
 
 import (
-	"archive/zip"
 	"bytes"
 	"fmt"
 	"io"
@@ -10,8 +9,6 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"os"
-	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 )
@@ -256,7 +253,7 @@ func TestDownload(t *testing.T) {
 	})
 
 	downloader.client = http
-	res, _ := downloader.Download(u)
+	res, _ := downloader.Download(u, false)
 
 	if string(res) != want {
 		t.Errorf("Response is %s, want %s", string(res), want)
@@ -281,7 +278,7 @@ func TestDownloadRetry(t *testing.T) {
 		callCount++
 	})
 	downloader.client = http
-	_, err := downloader.Download(u)
+	_, err := downloader.Download(u, false)
 	if err == nil {
 		t.Error("Expected error from downloader.Download(), got nil")
 	}
@@ -311,7 +308,7 @@ func TestDownloadWriteBack(t *testing.T) {
 	})
 
 	downloader.client = http
-	res, _ := downloader.Download(u)
+	res, _ := downloader.Download(u, false)
 
 	if string(res) != want {
 		t.Errorf("Response is %s, want %s", string(res), want)
@@ -352,7 +349,7 @@ func TestDownloadWriteBackSpecialFile(t *testing.T) {
 	})
 
 	downloader.client = http
-	res, _ := downloader.Download(u)
+	res, _ := downloader.Download(u, false)
 
 	if string(res) != want {
 		t.Errorf("Response is %s, want %s", string(res), want)
@@ -423,63 +420,4 @@ func TestRemoveRetry(t *testing.T) {
 	if callCount != 6 {
 		t.Errorf("Expected 6 retries, got %d", callCount)
 	}
-}
-
-// Taken from https://golangcode.com/unzip-files-in-go/
-func Unzip(src string, dest string) ([]string, error) {
-	var filenames []string
-
-	r, err := zip.OpenReader(src)
-	if err != nil {
-		return filenames, err
-	}
-	defer r.Close()
-
-	for _, f := range r.File {
-
-		rc, err := f.Open()
-		if err != nil {
-			return filenames, err
-		}
-		defer rc.Close()
-
-		// Store filename/path for returning and using later on
-		fpath := filepath.Join(dest, f.Name)
-
-		// Check for ZipSlip. More Info: http://bit.ly/2MsjAWE
-		if !strings.HasPrefix(fpath, filepath.Clean(dest)+string(os.PathSeparator)) {
-			return filenames, fmt.Errorf("%s: illegal file path", fpath)
-		}
-
-		filenames = append(filenames, fpath)
-
-		if f.FileInfo().IsDir() {
-
-			// Make Folder
-			os.MkdirAll(fpath, os.ModePerm)
-
-		} else {
-
-			// Make File
-			if err = os.MkdirAll(filepath.Dir(fpath), os.ModePerm); err != nil {
-				return filenames, err
-			}
-
-			outFile, err := os.OpenFile(fpath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
-			if err != nil {
-				return filenames, err
-			}
-
-			_, err = io.Copy(outFile, rc)
-
-			// Close the file without defer to close before next iteration of loop
-			outFile.Close()
-
-			if err != nil {
-				return filenames, err
-			}
-
-		}
-	}
-	return filenames, nil
 }
